\name{pubBonneryBreidtCoquet2016-package}
\alias{pubBonneryBreidtCoquet2016-package}
\alias{pubBonneryBreidtCoquet2016}
\docType{package}
\title{
\packageTitle{test}
}
\description{
\packageDescription{test}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pubBonneryBreidtCoquet2016}
\packageIndices{pubBonneryBreidtCoquet2016}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{pubBonneryBreidtCoquet2016}

Maintainer: \packageMaintainer{pubBonneryBreidtCoquet2016}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
